+++ lib/ndef.dart
@@ -1,62 +1,62 @@
 library ndef;
 
 // base class
-export 'src/record.dart';
+export 'record.dart';
 // utility
-export 'src/byteStream.dart';
+export 'utilities.dart';
 // record types
-export 'src/record/absoluteUri.dart';
-export 'src/record/mime.dart';
-export 'src/record/deviceinfo.dart';
-export 'src/record/signature.dart';
-export 'src/record/smartposter.dart';
-export 'src/record/text.dart';
-export 'src/record/uri.dart';
+export 'record/absoluteUri.dart';
+export 'record/mime.dart';
+export 'record/deviceinfo.dart';
+export 'record/handover.dart';
+export 'record/signature.dart';
+export 'record/smartposter.dart';
+export 'record/text.dart';
+export 'record/uri.dart';
 
 import 'dart:typed_data';
-import 'src/record.dart';
-import 'src/byteStream.dart';
+import 'record.dart';
+import 'utilities.dart';
 
-/// decode raw NDEF messages from byte array
-List<Record> decodeRawNdefMessage(Uint8List data,
-    {var typeFactory = Record.typeFactory}) {
-  var records = new List<Record>();
+/// Decode raw NDEF messages (containing at least one [NDEFRecord]) from byte array.
+List<NDEFRecord> decodeRawNdefMessage(Uint8List data,
+    {var typeFactory = NDEFRecord.defaultTypeFactory}) {
+  var records = new List<NDEFRecord>();
   var stream = new ByteStream(data);
   while (!stream.isEnd()) {
-    var record = Record.decodeStream(stream, typeFactory);
+    var record = NDEFRecord.decodeStream(stream, typeFactory);
     if (records.length == 0) {
-      if (record.flags.MB == false) {
-        throw "MB flag is not set in first record";
-      }
+      assert(record.flags.MB == true,"MB flag is not set in first record");
     } else {
-      if (record.flags.MB == true) {
-        throw "MB flag is set in middle record";
-      }
+      assert(record.flags.MB == false,"MB flag is set in middle record");
     }
     records.add(record);
   }
-  if (records.last.flags.ME != true) {
-    throw "ME flag is not set in last record";
-  }
-  if (records.last.flags.CF == true) {
-    throw "CF flag is set in last record";
-  }
+  assert(records.last.flags.ME == true,"ME flag is not set in last record");
+  assert(records.last.flags.CF == false,"CF flag is set in last record");
   return records;
 }
 
-/// decode partially parsed NDEF record
-Record decodePartialNdefMessage(
+/// Decode a NDEF record, providing its parts separately.
+/// This is most useful in mobile envrionment because the APIs will give you these information in a separate manner.
+NDEFRecord decodePartialNdefMessage(
     TypeNameFormat tnf, Uint8List type, Uint8List payload,
     {Uint8List id}) {
-  var decoded = Record.doDecode(tnf, type, payload, id: id);
+  var decoded = NDEFRecord.doDecode(tnf, type, payload, id: id);
   return decoded;
 }
 
-/// encode an NDEF message to byte array
-Uint8List encodeNdefMessage(List<Record> records, {bool canonicalize = true}) {
+/// Encode an NDEF message (containing several [NDEFRecord]s) to byte array.
+/// Set [canonicalize] to set the MB and ME fields automatically in the first / last record.
+Uint8List encodeNdefMessage(List<NDEFRecord> records,
+    {bool canonicalize = true}) {
   if (records.length == 0) {
     return new Uint8List(0);
   }
+
+  records.forEach((r) {
+    r.flags.resetPositionFlag();
+   });
 
   if (canonicalize) {
     records.first.flags.MB = true;