+++ lib/record/mime.dart
@@ -1,9 +1,8 @@
 import 'dart:typed_data';
 
-import '../record.dart';
-import '../byteStream.dart';
+import '../ndef.dart';
 
-class MimeRecord extends Record {
+class MimeRecord extends NDEFRecord {
   static const TypeNameFormat classTnf = TypeNameFormat.media;
 
   TypeNameFormat get tnf {
@@ -21,8 +20,7 @@ class MimeRecord extends Record {
   String toString() {
     var str = "MimeRecord: ";
     str += basicInfoString;
-    str += "type=$decodedType ";
-    str += "payload=${ByteStream.list2hexString(payload)}";
+    str += "payload=${ByteUtils.list2hexString(payload)}";
     return str;
   }
 }