+++ CHANGELOG.md
@@ -1,3 +1,14 @@
-## 0.1.0 [Unreleased]
+## 0.1.0
 
-* TODO: Describe initial release.
+* Initial release, preliminary decoding & encoding support
+* Support types:
+  * NFC Well Known
+    * Text
+    * URI with well-known prefix
+    * Digital signature
+    * Smart poster
+    * Connection handover
+  * Media (MIME data)
+    * Bluetooth easy pairing / Bluetooth low energy
+    * other MIME data
+  * Absolute URI